Moves PillButtonStyle to its own file
add pressed state to PillButton
